requirements.txt
Django==4.1.13
environs~=11.0.0
djangorestframework==3.11.2
psycopg2-binary==2.9.9
pre-commit==3.7.1
django-cors-headers==4.3.1
drf-yasg==1.21.7
uvicorn==0.29.0